+++ src/lib/trpc/routes/users.ts
@@ -145,10 +145,14 @@ export const usersRouter = t.router({
 						where: { magicLink: await hash(req.ctx.magicLink) },
 						data: { status: Status.CONFIRMED },
 					});
-
-					// notify user through email on confirming there RSVP
-					const subject = 'Thanks for Confirming!';
-					await sendEmail(req.ctx.user.email, subject, (await getSettings()).RSVPTemplate, null);
+					await sendEmail(
+						req.ctx.user.email,
+						'Thanks for your RSVP!',
+						(
+							await getSettings()
+						).confirmTemplate,
+						null
+					);
 				}
 			} else {
 				// Hackers should be able to decline after accepting and/or the deadline
@@ -157,15 +161,12 @@ export const usersRouter = t.router({
 						where: { magicLink: await hash(req.ctx.magicLink) },
 						data: { status: Status.DECLINED },
 					});
-
-					// notify user through email on successful withdrawal
-					const subject = 'Application Withdrawal Confirmation';
 					await sendEmail(
 						req.ctx.user.email,
-						subject,
+						'Thanks for your RSVP!',
 						(
 							await getSettings()
-						).withdrawTemplate,
+						).declineTemplate,
 						null
 					);
 				}
@@ -334,5 +335,67 @@ export const usersRouter = t.router({
 				throw new Error('You have insufficient permissions to perform this action.');
 			}
 			return await prisma.user.findMany({ orderBy: [{ id: 'asc' }], include: { decision: true } });
+		}),
+
+	/**
+	 * Bulk sets the status of all the users. User must be an admin.
+	 */
+	setStatuses: t.procedure
+		.use(authenticate)
+		.input(
+			z.object({
+				status: z.nativeEnum(Status),
+				ids: z.array(z.number()),
+			})
+		)
+		.mutation(async (req): Promise<void> => {
+			if (req.ctx.user.role !== Role.ADMIN) {
+				throw new Error('You have insufficient permissions to perform this action.');
+			}
+			const updateStatuses = prisma.user.updateMany({
+				where: {
+					id: {
+						in: req.input.ids,
+					},
+				},
+				data: {
+					status: req.input.status,
+				},
+			});
+			const deleteDecisions = prisma.decision.deleteMany({
+				where: {
+					userId: {
+						in: req.input.ids,
+					},
+				},
+			});
+			await prisma.$transaction([updateStatuses, deleteDecisions]);
+		}),
+
+	/**
+	 * Bulk sets the roles of all the users. User must be an admin.
+	 */
+	setRoles: t.procedure
+		.use(authenticate)
+		.input(
+			z.object({
+				role: z.nativeEnum(Role),
+				ids: z.array(z.number()),
+			})
+		)
+		.mutation(async (req): Promise<void> => {
+			if (req.ctx.user.role !== Role.ADMIN) {
+				throw new Error('You have insufficient permissions to perform this action.');
+			}
+			await prisma.user.updateMany({
+				where: {
+					id: {
+						in: req.input.ids,
+					},
+				},
+				data: {
+					role: req.input.role,
+				},
+			});
 		}),
 });